requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==4.1.5
django-appconf==1.0.5
django-cryptography==1.1
django-phonenumber-field==7.0.2
Pillow==9.4.0
pycparser==2.21
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
